+++ HISTORY.md
@@ -10,6 +10,7 @@
   * Updated default PageType to `post`
   * Added method `isPost()`
   * Added method `ThemeLayout.getCompiledPug()`
+  * Removed `KuzPage.getPageOptionsFN()`
 
 
 0.2.8 / 28-09-2020